chore: refactor recents and multiple widget tests
